renaming config moved to the frame test started
