+++ Interfaces/Evaluation/IMatching.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 namespace Interfaces.Evaluation
 {
     // The matcher gives back to the evaluation the coherent data for the evaulation
-    public interface IMathing : IInitializable
+    public interface IMatching : IInitializable
     {
         /// <summary>
         /// Gives the names of those measurement datas which are related to the given specification name
